+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/StackExchangeRedis/StackExchangeRedisConstants.cs
@@ -1,8 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-#if NET
-
 namespace OpenTelemetry.AutoInstrumentation.Instrumentations.StackExchangeRedis;
 
 internal static class StackExchangeRedisConstants
@@ -25,5 +23,3 @@ internal static class StackExchangeRedisConstants
 
     private const string ServerTypeTypeName = "StackExchange.Redis.ServerType";
 }
-
-#endif